# Prime the server to improve startup performance