--- a/ga/developer/javaee7/Dockerfile
+++ b/ga/developer/javaee7/Dockerfile
@@ -28,4 +28,4 @@ RUN if [ ! -z $REPOSITORIES_PROPERTIES ]; then mkdir /opt/ibm/wlp/etc/ \
 
 COPY server.xml /config/
 
-RUN server start && server stop && rm -rf /output/resources/security/ && rm -rf /output/messaging && rm -rf /output/workarea
+RUN server start && server stop && rm -rf /output/resources/security/ && rm -rf /output/messaging