--- a/ga/developer/webProfile6/Dockerfile
+++ b/ga/developer/webProfile6/Dockerfile
@@ -27,4 +27,4 @@ RUN if [ ! -z $REPOSITORIES_PROPERTIES ]; then mkdir /opt/ibm/wlp/etc/ \
 
 COPY server.xml /config/
 
-RUN server start && server stop && rm -rf /output/resources/security/
+RUN server start && server stop && rm -rf /output/resources/security/ && rm -rf /output/workarea