--- a/test/verify.sh
+++ b/test/verify.sh
@@ -30,6 +30,23 @@ waitForServerStart()
    return 1
 }
 
+waitForServerStop()
+{
+   cid=$1
+   end=$((SECONDS+120))
+   while (( $SECONDS < $end ))
+   do
+      result=$(docker logs $cid 2>&1 | grep "CWWKE0036I" | wc -l)
+      if [ $result = 1 ]
+      then
+         return 0
+      fi
+   done
+
+   echo "Liberty failed to stop within a reasonable time"
+   return 1
+}
+
 testLibertyStarts()
 {
    cid=$(docker run -d $image)
@@ -70,6 +87,7 @@ testLibertyStops()
       exit 1
    fi
 
+   waitForServerStop $cid
    docker logs $cid | grep -iq "CWWKE0036I"
    if [ $? != 0 ]
    then
@@ -118,7 +136,7 @@ testFeatureList()
 {
    version=$(docker run --rm $image sh -c 'echo $LIBERTY_VERSION')
    echo "Checking features for $image against version $version"
-   
+
    case $tag in
      microProfile)
        YAML_KEY='microProfile1'
@@ -132,13 +150,13 @@ testFeatureList()
      *)
        YAML_KEY=$tag
    esac
-   
+
    LIBERTY_URL=$(wget -qO- https://public.dhe.ibm.com/ibmdl/export/pub/software/websphere/wasdev/downloads/wlp/index.yml | grep $version -A 7 | sed -n "s/\s*$YAML_KEY:\s//p" | tr -d '\r' | head -n 1)
    if [ -z $LIBERTY_URL ]; then
      echo "WARNING: download not found - unable to verify features"
-     return 
+     return
    fi
-   
+
    if [[ $LIBERTY_URL == *jar ]]; then
      required_features=$(docker run --rm ibmjava:8-jre-alpine sh -c \
        "wget -q $LIBERTY_URL -U UA-IBM-WebSphere-Liberty-Docker -O /tmp/wlp.jar
@@ -150,15 +168,15 @@ java -jar /tmp/wlp.jar --acceptLicense /opt/ibm > /dev/null
 unzip -q /tmp/wlp.zip -d /opt/ibm
 /opt/ibm/wlp/bin/productInfo featureInfo" | cut -d ' ' -f1 | sort)
    fi
-   
+
    actual_features=$(docker run --rm $image productInfo featureInfo | cut -d " " -f1 | sort)
-   
+
    additional_features=$(comm -1 -3 <(echo "$required_features") <(echo "$actual_features"))
    if [ "$additional_features" != "" ]; then
      echo "Additional features installed"
      echo "$additional_features"
    fi
-   
+
    missing_features=$(comm -2 -3 <(echo "$required_features") <(echo "$actual_features"))
    if [ "$missing_features" != "" ]; then
      echo "Missing features, exiting"
@@ -185,4 +203,3 @@ do
    eval $name
    echo "*** $name - Completed successfully"
 done
-