Adding a server start/stop to the images (except kernel, since there's no server.xml in there) to improve startup time.  There's a tradeoff with image sizes, as noted below:

microprofile:  increase of 63 MB.   startup from 8 seconds to 3 seconds
webProfile7:   increase of 64 MB.   startup from 9 seconds to 4 seconds
javaee7:      increase of  63 MB.   startup from 12 seconds to 5 seconds

Signed-off-by: Arthur De Magalhaes <ademagalhaes@gmail.com>