@arthurdm 2018-04-18
Thanks guys. Good points.  I pushed the changes to avoid removing the workarea after server priming (and remove the corresponding test for workarea).  

I will open an issue for the image size/startup display, that's a good idea.